Lab_Cod/requirements.txt
fastapi
uvicorn
scipy
opencv-python
matplotlib
pillow
PyWavelets
python-multipart
httpx